refactor(web): split DebugPanel and DashPanel into partner dirs
The last two long admin panels, same pattern as the three before them.

DebugPanel 1525 -> 329, with components/admin/debug/:
  types.ts           the shapes of the controller's /debug response
  format.ts          shared formatting helpers
  bits.tsx           health cells, kv tables, JSON blocks, call sections, chips
  BudgetMeter.tsx    the daily token-budget meter
  TtsPanels.tsx      TTS routing + recent calls
  SessionChat.tsx    the live DJ session as the chat history it is
  MountsTable.tsx    per-mount stream status
  DjContext.tsx      time / weather / festival / dominant mood
  FilesTable.tsx     the state-dir file listing
  LlmCalls.tsx       the recent-LLM-call log
  SubsonicCalls.tsx  the recent-Subsonic-call log

This is the biggest win of the set: DebugPanel was ~300 lines of panel
and ~1200 lines of eleven independent read-only cards, each of which is
now findable by name.

DashPanel 1292 -> 824, with components/admin/dash/:
  types.ts          shapes + the say/segment/sort vocabularies
  bits.tsx          sortable header, segment buttons, toggle rows, booth turn
  LikesCard.tsx     most-liked and most-recent likes
  RequestsCard.tsx  listener requests and one request's row

SettingsPanel (1350) is deliberately untouched: ~1290 of it is a single
component's state and handlers, with only ~25 lines of extractable
constants. Splitting it means decomposing the component, which is a
design change rather than a move, so it doesn't belong in this PR.

No behaviour change. Web lint clean and `next build` succeeds.
