+++ web/components/admin/dash/types.ts
@@ -1,0 +1,182 @@
+// Shapes and the fixed vocabularies the dashboard offers: say kinds and modes,
+// the manual segment buttons, and the listener-table sort keys.
+//
+// Part of the dash/ split - see ../DashPanel.tsx.
+
+import { clientLabel, type ListenerConnection } from '../../../lib/clientLabel';
+import type { SessionTurn } from '../../../lib/types';
+import type {
+  NowPlayingTrack,
+  StationContext,
+  ActiveShow,
+  DjState,
+  ListenerCount,
+  QueueEntry,
+  StationLocale,
+} from '../../../lib/types';
+import { AudioLines, Clock3, MessagesSquare, RadioTower, type LucideIcon } from 'lucide-react';
+
+
+export const SAY_KINDS = [
+  { id: 'dj-speak', label: 'Solo' },
+  { id: 'link', label: 'Over' },
+];
+export const SAY_MODES = [
+  { id: 'raw', label: 'Raw' },
+  { id: 'styled', label: 'Styled' },
+];
+
+// Canned prompts for the manual voice box, in station voice. Written as
+// instructions (they shine with mode=styled — the DJ rewrites them in persona,
+// though raw works too). Clicking one FILLS the textarea; nothing goes to air
+// until the operator hits send. This set is the zero-latency initial render
+// and the fallback when the controller has no generated batch — the ↻ button
+// swaps in fresh LLM-written ones via /generate/say-suggestions. The
+// controller keeps the canonical copy of these six as the generator's style
+// anchors (llm/internal/prompts/generate.ts SAY_SUGGESTION_EXAMPLES); keep
+// the two lists in step.
+export const SAY_SUGGESTIONS = [
+  'Tease the weather like it’s a rumour you can’t quite confirm.',
+  'Do a station ID like you suspect nobody’s listening — and you’re fine with it.',
+  'Salute the graveyard shift: night drivers, dish pits, the deliberately awake.',
+  'Tease the next track without giving up the title.',
+  'Remind everyone the request line exists and judges no one.',
+  'Announce the time like it’s classified information.',
+];
+
+type SegmentType = 'station-id' | 'hourly' | 'link' | 'banter';
+export const SEGMENTS: { type: SegmentType; label: string; icon: LucideIcon }[] = [
+  { type: 'station-id', label: 'Station ID', icon: RadioTower },
+  { type: 'hourly', label: 'Time check', icon: Clock3 },
+  { type: 'link', label: 'Track link', icon: AudioLines },
+];
+// Only offered while a show with guest co-hosts is on air — a one-person
+// "exchange" is a 400 from the controller anyway.
+export const BANTER_SEGMENT: { type: SegmentType; label: string; icon: LucideIcon } =
+  { type: 'banter', label: 'Banter', icon: MessagesSquare };
+
+export interface QueueState {
+  upcoming?: QueueEntry[];
+  history?: QueueEntry[];
+  autoPick?: boolean;
+  autoLink?: boolean;
+  pickerBusy?: boolean;
+}
+
+export interface DashStatus {
+  nowPlaying?: NowPlayingTrack | null;
+  context?: StationContext | null;
+  dj?: DjState | null;
+  listeners?: ListenerCount | number | null;
+  streamOnline?: boolean;
+  streamBitrate?: number | null;
+  activeShow?: ActiveShow | null;
+  queue?: QueueState;
+  sessionMessages?: SessionTurn[];
+  /** Station IANA zone — render on-air timestamps in it (issue #418). */
+  timezone?: string;
+  locale?: StationLocale;
+}
+
+// Subset of /stats (admin) the health strip reads: DJ p95 latency + the TTS
+// fallback rate, both since-boot rollups. Polled slower than live status since
+// they move slowly and the endpoint is heavier.
+export interface HealthStats {
+  llm?: { count?: number; latency?: { p95?: number }; agentTimeoutMs?: number };
+  tts?: { count?: number; fallbackRate?: number | null };
+}
+
+export interface ActResponse {
+  ok?: boolean;
+  spoken?: string;
+  error?: string;
+}
+
+export interface ConnectionsState {
+  count: number;
+  connections: ListenerConnection[];
+}
+
+// One resolved/failed listener request, as returned by GET /requests. Mirrors
+// the durable record written by the controller's request-log.
+export interface RequestEntry {
+  t?: string;
+  requester?: string;
+  text?: string;
+  status?: string;
+  ms?: number | null;
+  path?: string | null;
+  pickSource?: string | null;
+  intent?: string | null;
+  mood?: string | null;
+  scope?: string | null;
+  sort?: string | null;
+  artist?: string | null;
+  genre?: string | null;
+  language?: string | null;
+  searchTerms?: string[] | null;
+  track?: { title?: string; artist?: string; id?: string } | null;
+  ack?: string | null;
+  introScript?: string | null;
+  message?: string | null;
+}
+
+// Listener likes (#991), as returned by GET /likes — the heart button's
+// admin review surface.
+interface LikeTopEntry {
+  track?: { id?: string; title?: string; artist?: string } | null;
+  count?: number;
+  lastLikedAt?: string;
+}
+interface LikeRecentEntry {
+  songId?: string;
+  title?: string;
+  artist?: string;
+  album?: string;
+  likedAt?: string;
+  listener?: string;
+}
+export interface LikesPayload {
+  totals?: { total?: number; songs?: number };
+  top?: LikeTopEntry[];
+  recent?: LikeRecentEntry[];
+}
+
+// Hide the host portion of an IP so a glance at the screen doesn't expose a
+// listener's full address. IPv4 drops the last octet, IPv6 keeps the first two
+// groups (the routing prefix) and masks the rest. The raw IP is still in the
+// row's title attribute and one toggle away — this is a display default, not
+// redaction.
+export function maskIp(ip: string): string {
+  if (!ip) return '—';
+  if (ip.includes('.')) return ip.replace(/\.\d+$/, '.×');
+  if (ip.includes(':')) {
+    const groups = ip.split(':').filter(Boolean);
+    return groups.length > 2 ? `${groups[0]}:${groups[1]}:×` : ip;
+  }
+  return ip;
+}
+
+export type SortKey = 'ip' | 'mount' | 'connectedSeconds' | 'client';
+export interface SortState {
+  key: SortKey;
+  dir: 'asc' | 'desc';
+}
+
+// Sort connections by the active column. `client` sorts on the friendly label
+// (what the operator actually sees), everything else on the raw field.
+export function sortConnections(
+  rows: ListenerConnection[],
+  { key, dir }: SortState,
+): ListenerConnection[] {
+  const sign = dir === 'asc' ? 1 : -1;
+  return [...rows].sort((a, b) => {
+    let cmp: number;
+    if (key === 'connectedSeconds') cmp = a.connectedSeconds - b.connectedSeconds;
+    else if (key === 'client') cmp = clientLabel(a.userAgent).localeCompare(clientLabel(b.userAgent));
+    else cmp = String(a[key]).localeCompare(String(b[key]));
+    return cmp * sign;
+  });
+}
+
+
