+++ web/components/admin/debug/types.ts
@@ -1,0 +1,241 @@
+// The shapes of the controller's /debug response.
+//
+// All admin endpoints return loose JSON, so these are narrowed with
+// optional-chaining at the call sites rather than trusted outright.
+//
+// Part of the debug/ split - see ../DebugPanel.tsx.
+
+import type { StationLocale } from '../../../lib/types';
+
+// All admin endpoints return loose JSON; type as unknown then narrow with
+// optional-chaining at call sites. The shapes mirror the controller's
+// /debug response.
+export interface DebugIcecast {
+  listeners?: number;
+  peakListeners?: number;
+  error?: string;
+}
+
+interface DebugLibrary {
+  total?: number;
+  updatedAt?: string;
+}
+
+interface DebugQueueEntry {
+  title?: string;
+  artist?: string;
+  requestedBy?: string;
+}
+
+interface DjLogEntry {
+  id: string;
+  t?: string;
+  kind?: string;
+  message?: string;
+}
+
+interface DebugQueue {
+  current?: Record<string, unknown> | null;
+  upcoming?: DebugQueueEntry[];
+  djLog?: DjLogEntry[];
+  djLogCount?: number;
+}
+
+interface DebugTtsSpoken {
+  engine?: string;
+  voice?: string;
+  provider?: string;
+  fellBack?: boolean;
+  requested?: string;
+}
+
+/** One entry from the controller's TTS call ring (stats.ts ttsCalls) — every
+ * speak() outcome since boot, newest first, incl. silent engine fallbacks. */
+export interface TtsCall {
+  ok?: boolean;
+  kind?: string;
+  engine?: string;
+  requested?: string;
+  fellBack?: boolean;
+  ms?: number;
+  chars?: number;
+  t?: string;
+  error?: string;
+  /** Spoken text, capped at ~240 chars by the controller. */
+  text?: string;
+  /** Voicing persona name; null for global kinds (jingle/default). */
+  persona?: string | null;
+}
+
+export interface DebugTts {
+  spoken?: DebugTtsSpoken;
+  jingle?: { engine?: string };
+  effectivePersona?: { name?: string };
+  recentCalls?: TtsCall[];
+  error?: string;
+}
+
+interface SessionMessage {
+  t?: string;
+  role?: string;
+  kind?: string;
+  text?: string;
+}
+
+export interface DebugSession {
+  kind?: string;
+  show?: { name?: string };
+  persona?: { name?: string };
+  messages?: SessionMessage[];
+  handoff?: string;
+  error?: string;
+}
+
+interface LlmCall {
+  ok?: boolean;
+  kind?: string;
+  model?: string;
+  via?: string;
+  ms?: number;
+  t?: string;
+  error?: string;
+  user?: string;
+  system?: string;
+  systemPreview?: string;
+  messages?: Array<{ role?: string; content?: unknown }>;
+  toolCalls?: Array<{ name?: string; args?: unknown; result?: unknown }>;
+  response?: string;
+  /** What the model said INSTEAD of the expected structured output on a
+   * failed call (e.g. "agent did not call the done tool") — one labelled
+   * block per attempt that declined. Populated by failureDiagnostics() in
+   * the controller; absent on success (see `response` instead). */
+  responseText?: string;
+  steps?: number;
+}
+
+export interface DebugLlm {
+  activeModel?: string;
+  provider?: string;
+  budget?: DebugBudget;
+  recentCalls?: LlmCall[];
+  /** Raw-request capture status — drives the toggle + file-path hint. */
+  debug?: {
+    enabled?: boolean;
+    viaEnv?: boolean;
+    file?: string;
+    max?: number;
+  };
+}
+
+interface SubsonicEndpoint {
+  endpoint: string;
+  calls: number;
+}
+
+interface SubsonicCall {
+  ok?: boolean;
+  endpoint?: string;
+  count?: number;
+  ms?: number;
+  t?: string;
+  error?: string;
+  params?: Record<string, unknown>;
+  songIds?: Array<{ title?: string; artist?: string }>;
+}
+
+export interface DebugSubsonic {
+  recentCalls?: SubsonicCall[];
+  endpoints?: SubsonicEndpoint[];
+  error?: string;
+}
+
+interface FileEntry {
+  name: string;
+  size?: number;
+  mtime?: string;
+  isDir?: boolean;
+}
+
+export type FilesValue = FileEntry[] | { error?: string } | undefined;
+
+// Mirrors the controller's getFullContext() snapshot — what the DJ "feels"
+// right now. Rendered human-friendly by <DjContext>, not as raw JSON.
+export interface DebugContext {
+  time?: { period?: string; mood?: string; vibe?: string; show?: string };
+  weather?: {
+    condition?: string;
+    mood?: string | null;
+    temp?: number | null;
+    tempUnit?: string;
+    isDay?: boolean;
+    location?: string;
+  };
+  festival?: { name?: string; mood?: string } | null;
+  dominantMood?: string;
+  date?: {
+    iso?: string;
+    dayOfWeek?: number;
+    dayLabel?: string;
+    monthLabel?: string;
+    dayOfMonth?: number;
+    season?: string;
+  };
+  clock?: { hhmm?: string; isWeekend?: boolean; isLateNight?: boolean; isCommute?: boolean };
+  activeShow?: { name?: string; mood?: string; topic?: string; persona?: { name?: string } | null } | null;
+  listeners?: { count?: number | null };
+  error?: string;
+}
+
+export interface DebugMount {
+  path: string;
+  codec: string;
+  configured: boolean;
+  live: boolean;
+  bitrate: number | null;
+  listeners: number | null;
+  sampleRate: number | null;
+  channels: number | null;
+  contentType: string | null;
+  url: string;
+}
+
+export interface DebugMounts {
+  list: DebugMount[];
+  tuneIn: { entryCount: number; pls: string; m3u: string };
+}
+
+/** Daily token budget snapshot (settings.llm.dailyTokenCap) — mirrors the
+ * controller's budgetMode() tiers: soft mutes optional segments, hard stops
+ * model calls entirely until the UTC day rolls. */
+export interface DebugBudget {
+  enabled?: boolean;
+  cap?: number;
+  softPct?: number;
+  exemptRequests?: boolean;
+  usedToday?: number;
+  remaining?: number;
+  mode?: 'normal' | 'soft' | 'hard';
+}
+
+export interface DebugData {
+  /** Station IANA zone — render DJ-log timestamps in it (issue #418). */
+  timezone?: string;
+  locale?: StationLocale;
+  icecast?: DebugIcecast;
+  liquidsoapLog?: string;
+  llm?: DebugLlm;
+  queue?: DebugQueue;
+  library?: DebugLibrary;
+  nowPlaying?: Record<string, unknown> | null;
+  context?: DebugContext | null;
+  tts?: DebugTts;
+  subsonic?: DebugSubsonic;
+  session?: DebugSession;
+  stateFiles?: FilesValue;
+  voiceFiles?: FilesValue;
+  config?: Record<string, unknown>;
+  mounts?: DebugMounts;
+  error?: string;
+}
+
+
